Support parent_name and tag in binding:profile
OVN has some special support defined for containers running inside VMs
that are connected to networks managed by OVN.  You can create ports
for these containers that have a parent port (the VM) associated with
them.  You also set a tag that lets OVN differentiate traffic from the
containers from the traffic from the VM.  This lets you connect those
containers directly to any network(s) managed by Neutron, without
necessarily having the VM attached to the same network(s).

This concept for parent and child ports is not something that exists
in Neutron today.  To initially test and prove out the concept, we
make use of 'binding:profile', which allows arbitrary backend specific
data to be passed in through the API.  If this works out well, we
should revisit making this an official extension of the Neutron API.

Change-Id: Iade522f0f145ca33ee96c40be05eb7198cce9415